+++ src/main/java/step03/domain/Steps.java
@@ -1,4 +1,4 @@
-package step03.domain.domain;
+package step03.domain;
 
 import exception.InvalidCountOfStepsException;
