+++ src/test/java/step03/ParticipantsTest.java
@@ -1,4 +1,4 @@
-package step03.domain;
+package step03;
 
 import exception.DuplicatedParticipantsNameException;
 import exception.InvalidCountOfParticipantsException;
