+++ src/test/java/step03/PointTest.java
@@ -1,4 +1,4 @@
-package step03.domain;
+package step03;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
